kubepushnotif/requirements.txt
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
futures==3.2.0
tornado
pika==0.12.0
python-memcached-stats
pymemcache
singledispatch
backports_abc
six
websocket-client
multi-mechanize
matplotlib